Handle missing argument exception
