+++ src/main/java/Ponder_Pika/Task/Deadline.java
@@ -1,4 +1,4 @@
-package Task;
+package Ponder_Pika.Task;
 
 public class Deadline extends Task {
     private final String deadline;
